+++ lib/src/customizer_scope_adder.dart
@@ -46,11 +46,11 @@ class CustomizerScopeAdder {
 
   // ...........................................................................
   /// Deeply iterate through all child nodes and replace nodes
-  void init(Scope scope) {
-    _initScope(scope);
+  void applyToScope(Scope scope) {
+    _applyToScope(scope);
 
     for (final childScope in scope.children) {
-      init(childScope);
+      applyToScope(childScope);
     }
   }
 
@@ -59,7 +59,7 @@ class CustomizerScopeAdder {
   // ######################
 
   // ...........................................................................
-  void _initScope(Scope scope) {
+  void _applyToScope(Scope scope) {
     final bluePrints = customizer.bluePrint.addScopes(
       hostScope: scope,
     );