+++ test/customizer_test.dart
@@ -15,6 +15,39 @@ void main() {
       test('should add itself the the host scope', () {
         expect(customizer.scope.customizers, contains(customizer));
       });
+
+      test('should be applied to nodes added after instantiation', () {
+        final customizer = ExampleCustomizerBluePrint.example;
+
+        // Make sure, example customizer is added to the scope
+        final scope = customizer.scope;
+        expect(scope.customizers.first, customizer);
+
+        // The customizer applied inserts
+        final hostA = scope.findNode<int>('hostA')!;
+        final hostB = scope.findNode<int>('hostB')!;
+        final hostC = scope.findNode<int>('hostC')!;
+
+        expect(hostA.inserts, hasLength(2));
+        expect(hostB.inserts, hasLength(3));
+        expect(hostC.inserts, hasLength(3));
+
+        // Let's add two more nodes to the scopeA and scopeB
+        final scopeA = scope.findScope('a')!;
+        final scopeB = scope.findScope('b')!;
+
+        final hostA1 =
+            const NodeBluePrint<int>(key: 'hostA1', initialProduct: 11)
+                .instantiate(scope: scopeA);
+
+        final hostB1 =
+            const NodeBluePrint<int>(key: 'hostB1', initialProduct: 12)
+                .instantiate(scope: scopeB);
+
+        // The customizers are applied to the newly added nodes
+        expect(hostA1.inserts, hasLength(2));
+        expect(hostB1.inserts, hasLength(3));
+      });
     });
 
     group('dispose', () {