+++ test/node_blue_print_test.dart
@@ -269,6 +269,15 @@ void main() {
         expect(newBluePrint.suppliers, bluePrint.suppliers);
         expect(newBluePrint.produce([], 0), 1);
       });
+
+      test('should apply customizers', () {
+        final customizer = Customizer.example();
+        final scope = customizer.scope;
+        final newNode =
+            const NodeBluePrint<int>(key: 'hostX', initialProduct: 0)
+                .instantiate(scope: scope);
+        expect(newNode.inserts, isNotEmpty);
+      });
     });
 
     group('forwardTo(key)', () {